requirements.txt
## RecBole Dependencies
recbole
numpy>=1.17.2
scipy>=1.6.0
hyperopt==0.2.5
pandas>=1.3.0
tqdm>=4.48.2
scikit_learn>=0.23.2
pyyaml>=5.1.0
colorlog>=4.7.2
colorama>=0.4.4
tensorboard>=2.5.0
thop>=0.1.1.post2207130030
ray>=1.13.0, <=2.6.3
tabulate>=0.8.10 
plotly>=4.0.0
texttable>=0.9.0
psutil>=5.9.0
kmeans-pytorch
##
## LLMs
openai
tiktoken
transformers==4.34.1
bitsandbytes
datasets
accelerate
peft
trl
##
## General
pandas
numpy
python-dotenv
wandb
